+++ homepare/src/App.jsx
@@ -12,7 +12,7 @@ import { Checklist } from './checklist'
 import { CollectionDetail } from './CollectionDetail'
 import { ComparisonTable } from './comparisonTable'
 import { SearchDetailsCard } from './searchDetailsCard'
-import { ListingDetailsCard } from './ListingDetails'
+import { ListingDetailsCard } from './ListingDetailsCard'
 import { UserAccount } from './UserAccount'
 import { UserListings } from './UserListings'
 import { UserCollections } from './UserCollections'
