+++ homepare/src/ListingDetailsCard.jsx
